- Fixing namespaces - Moving web server setup to parent class - Making which application to run configurable on a class level
