+++ Fhi.HelseId.Integration.Tests/IntegrationTest.cs
@@ -3,14 +3,13 @@ using System.Net.Http.Headers;
 using System.Text.Json;
 using Microsoft.AspNetCore.Mvc.Testing;
 
-namespace Fhi.HelseId.Integration.Tests
+namespace Fhi.HelseId.Integration.Tests.Setup
 {
     [TestFixture]
-    public abstract class IntegrationTest
+    public abstract class IntegrationTest<TProgram> where TProgram : class
     {
-        internal WebApplicationFactory<Program>? _factory;
+        public readonly WebApplicationFactory<TProgram> Factory = new();
         internal Dictionary<TokenType, String> _tokens = new();
-        public WebApplicationFactory<Program> Factory => _factory!;
 
         [OneTimeSetUp]
         public async Task CreateTokens()
@@ -26,14 +25,10 @@ namespace Fhi.HelseId.Integration.Tests
             }
         }
 
+        
         public static string GetDirectoryForCaller(
             [System.Runtime.CompilerServices.CallerFilePath] string sourceFilePath = ""
         ) => sourceFilePath[..sourceFilePath.LastIndexOf('\\')];
-
-        public void CreateService()
-        {
-            _factory = new();
-        }
 
         public HttpClient CreateHttpClient(TokenType tokenType)
         {
