+++ Fhi.HelseId.Integration.Tests/Setup/Tokens.cs
@@ -1,6 +1,6 @@
-using Fhi.HelseId.Integration.Tests.TttClient;
+using Fhi.HelseId.Integration.Tests.Setup.TttClient;
 
-namespace Fhi.HelseId.Integration.Tests;
+namespace Fhi.HelseId.Integration.Tests.Setup;
 
 public enum TokenType
 {
