+++ Fhi.HelseId.Integration.Tests/Setup/TttClient/TttClient.cs
@@ -18,7 +18,7 @@
 #pragma warning disable 8625 // Disable "CS8625 Cannot convert null literal to non-nullable reference type"
 #pragma warning disable 8765 // Disable "CS8765 Nullability of type of parameter doesn't match overridden member (possibly because of nullability attributes)."
 
-namespace Fhi.HelseId.Integration.Tests.TttClient
+namespace Fhi.HelseId.Integration.Tests.Setup.TttClient
 {
     using System = global::System;
